+++ src/effect/effecthandler.cpp
@@ -655,7 +655,7 @@ bool EffectsHandler::tabletPadStripEvent(int number, qreal position, bool isFing
     return false;
 }
 
-bool EffectsHandler::tabletPadRingEvent(int number, int position, bool isFinger, std::chrono::microseconds time, InputDevice *device)
+bool EffectsHandler::tabletPadRingEvent(int number, qreal position, bool isFinger, std::chrono::microseconds time, InputDevice *device)
 {
     // TODO: reverse call order?
     for (auto it = loaded_effects.constBegin(); it != loaded_effects.constEnd(); ++it) {
